requirements.txt
grpcio==1.30.0
grpcio-tools==1.30.0
gensim==3.8.0
spacy==2.3.2